로그인 기능 구현

+++ src/main/resources/static/js/list2.js
@@ -2,6 +2,8 @@ window.addEventListener('DOMContentLoaded', () => {
     getGuestbooks();
 });
 
+curStart = 0;
+
 showLoadingText = (show) => {
     let loadingElement = document.querySelector(".loading");
     let containerElement = document.querySelector(".container");
@@ -14,6 +16,14 @@ getGuestbooks = (start = 0) => {
     showLoadingText(true);
 
     fetchData(`/guestbooks?start=${start}`, null, "GET", json => {
+        if (start > 0 && json.list.length == 0) {
+            getGuestbooks(0);
+            return;
+        }
+
+        let isAdmin = document.querySelector("#isAdmin").dataset.isAdmin;
+        console.log("isAdminTag: " + document.querySelector("#isAdmin"));
+        console.log('isAdmin : ' + isAdmin);
         document.querySelector(".count").innerText = json.count
         let listHtml = json.list.reduce((accumulator, guestbook) => {
             accumulator +=
@@ -24,14 +34,19 @@ getGuestbooks = (start = 0) => {
                     내용: ${guestbook.content}<br>
                     작성일: ${guestbook.regdate}
                 </div>
-                <input 
-                    type="button" 
-                    onclick="deleteGuestbook(${guestbook.id})" 
-                    value="삭제">
-            </li>`;
+            `;
+
+            if (isAdmin) {
+                accumulator += ` <input 
+                type="button" 
+                onclick="deleteGuestbook(${guestbook.id})" 
+                value="삭제">`
+            }
+            accumulator += "</li>"
             return accumulator;
         }, "");
         document.querySelector(".list").innerHTML = listHtml;
+
 
         let pagesHtml = json.pageStartList.reduce((accumulator, pageStart, currentIndex) => {
             accumulator +=
@@ -45,13 +60,13 @@ getGuestbooks = (start = 0) => {
         document.querySelector(".pages").innerHTML = pagesHtml;
 
         let welcomeH2Tag = document.querySelector(".welcome")
-        if(Boolean(json.firstVisit)){
-            welcomeH2Tag.innerText="처음 오셨군요. 환영합니다!";
+        if (Boolean(json.firstVisit)) {
+            welcomeH2Tag.innerText = "처음 오셨군요. 환영합니다!";
         }
-        else{
-            welcomeH2Tag.innerText="또 오셨군요. 환영합니다!";
+        else {
+            welcomeH2Tag.innerText = "또 오셨군요. 환영합니다!";
         }
-
+        curStart = start;
         showLoadingText(false);
     });
 };
@@ -59,7 +74,7 @@ deleteGuestbook = (guest_id) => {
     showLoadingText(true);
     fetchData(`/guestbooks/${guest_id}`, null, "DELETE", json => {
         if (json.success) {
-            getGuestbooks();
+            getGuestbooks(curStart);
         }
         else {
             alert('삭제 실패!');
@@ -73,18 +88,18 @@ addGuestbook = (submitTag) => {
     let submitForm = submitTag.form;
     let name = submitForm.name.value;
     let content = submitForm.content.value;
-    if(!name || !content) {
+    if (!name || !content) {
         alert('이름 또는 내용을 입력해주세요!');
         return;
     }
     fetchData(
-        "/guestbooks", 
-        {name, content}, 
+        "/guestbooks",
+        { name, content },
         "POST",
-        json=>{
-            if(json && json.id && json.name){
+        json => {
+            if (json && json.id && json.name) {
                 getGuestbooks();
-            }else{
+            } else {
                 alert("방명록 작성 실패!")
             }
         }
